+++ node/core/containerCaller_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/opctl/opctl/util/pubsub"
 	"github.com/opspec-io/sdk-golang/model"
 	"github.com/pkg/errors"
+	"sync"
 	"time"
 )
 
@@ -62,7 +63,7 @@ var _ = Context("containerCaller", func() {
 			/* assert */
 			Expect(fakeDCGNodeRepo.AddArgsForCall(0)).To(Equal(expectedDCGNodeDescriptor))
 		})
-		It("should call pubSub.Publish w/ expected args", func() {
+		It("should call pubSub.Publish w/ expected ContainerStartedEvent", func() {
 			/* arrange */
 			providedInboundScope := map[string]*model.Data{}
 			providedContainerId := "dummyContainerId"
@@ -219,7 +220,122 @@ var _ = Context("containerCaller", func() {
 		/* assert */
 		Expect(fakeDCGNodeRepo.DeleteIfExistsArgsForCall(0)).To(Equal(providedContainerId))
 	})
-	It("should call pubSub.Publish w/ expected args", func() {
+
+	It("should call pubSub.Publish w/ expected ContainerExitedEvent", func() {
+		/* arrange */
+		providedInboundScope := map[string]*model.Data{}
+		providedContainerId := "dummyContainerId"
+		providedScgContainerCall := &model.ScgContainerCall{}
+		providedPkgRef := "dummyPkgRef"
+		providedRootOpId := "dummyRootOpId"
+
+		expectedEvent := &model.Event{
+			Timestamp: time.Now().UTC(),
+			ContainerExited: &model.ContainerExitedEvent{
+				ContainerId: providedContainerId,
+				PkgRef:      providedPkgRef,
+				RootOpId:    providedRootOpId,
+			},
+		}
+
+		fakePubSub := new(pubsub.Fake)
+		fakePubSub.SubscribeStub = func(filter *model.EventFilter, eventChannel chan *model.Event) {
+			close(eventChannel)
+		}
+
+		objectUnderTest := newContainerCaller(
+			new(containerprovider.Fake),
+			fakePubSub,
+			new(fakeDCGNodeRepo),
+		)
+
+		/* act */
+		objectUnderTest.Call(
+			providedInboundScope,
+			providedContainerId,
+			providedScgContainerCall,
+			providedPkgRef,
+			providedRootOpId,
+		)
+
+		/* assert */
+		actualEvent := fakePubSub.PublishArgsForCall(1)
+
+		// @TODO: implement/use VTime (similar to VOS & VFS) so we don't need custom assertions on temporal fields
+		Expect(actualEvent.Timestamp).To(BeTemporally("~", time.Now().UTC(), 5*time.Second))
+		// set temporal fields to expected vals since they're already asserted
+		actualEvent.Timestamp = expectedEvent.Timestamp
+
+		Expect(actualEvent).To(Equal(expectedEvent))
+	})
+	XIt("should call pubSub.Publish w/ expected OutputInitializedEvents", func() {
+		/* arrange */
+		providedInboundScope := map[string]*model.Data{}
+		providedContainerId := "dummyContainerId"
+		providedScgContainerCall := &model.ScgContainerCall{
+			Sockets: map[string]string{
+				"0.0.0.0": "socket0Name",
+			},
+			Files:  map[string]string{},
+			Dirs:   map[string]string{},
+			StdErr: map[string]string{},
+			StdOut: map[string]string{},
+		}
+		providedPkgRef := "dummyPkgRef"
+		providedRootOpId := "dummyRootOpId"
+
+		expectedEventTimestamp := time.Now().UTC()
+
+		expectedOutputInitEvents := []*model.Event{
+			{
+				Timestamp: expectedEventTimestamp,
+				OutputInitialized: &model.OutputInitializedEvent{
+					Name:     providedScgContainerCall.Sockets["0.0.0.0"],
+					Value:    &model.Data{Socket: providedContainerId},
+					RootOpId: providedRootOpId,
+					CallId:   providedContainerId,
+				},
+			},
+		}
+
+		fakePubSub := new(pubsub.Fake)
+
+		// record actual published events
+		var actualOutputInitEventMutex sync.Mutex
+		actualOutputInitEvents := []*model.Event{}
+		fakePubSub.PublishStub = func(event *model.Event) {
+			event.Timestamp = expectedEventTimestamp
+			actualOutputInitEventMutex.Lock()
+			actualOutputInitEvents = append(actualOutputInitEvents, event)
+			actualOutputInitEventMutex.Unlock()
+		}
+		fakePubSub.SubscribeStub = func(filter *model.EventFilter, eventChannel chan *model.Event) {
+			close(eventChannel)
+		}
+
+		objectUnderTest := newContainerCaller(
+			new(containerprovider.Fake),
+			fakePubSub,
+			new(fakeDCGNodeRepo),
+		)
+
+		/* act */
+		objectUnderTest.Call(
+			providedInboundScope,
+			providedContainerId,
+			providedScgContainerCall,
+			providedPkgRef,
+			providedRootOpId,
+		)
+
+		/* assert */
+		for _, expectedOutputInitEvent := range expectedOutputInitEvents {
+			Eventually(func() []*model.Event {
+				return actualOutputInitEvents
+			}, 5*time.Second, 1*time.Second).Should(ContainElement(expectedOutputInitEvent))
+		}
+	})
+	It("should call pubSub.Publish w/ expected ContainerExitedEvent", func() {
 		/* arrange */
 		providedInboundScope := map[string]*model.Data{}
 		providedContainerId := "dummyContainerId"