+++ node/node.go
@@ -8,7 +8,7 @@ import (
 	"github.com/opctl/opctl/util/containerprovider/docker"
 	"github.com/opctl/opctl/util/lockfile"
 	"github.com/opctl/opctl/util/pubsub"
-	"github.com/opctl/opctl/util/vfs/os"
+	"github.com/virtual-go/vfs/osfs"
 	"path"
 )
 
@@ -30,15 +30,8 @@ func New() {
 	// cleanup [legacy] opspec.engine container if exists; ignore errors
 	containerProvider.DeleteContainerIfExists("opspec.engine")
 
-	// (re)create network; ignore errors
-	containerProvider.DeleteNetworkIfExists("opctl")
-	createNetworkErr := containerProvider.CreateNetwork("opctl")
-	if nil != createNetworkErr {
-		fmt.Print(createNetworkErr.Error())
-	}
-
 	// cleanup existing DCG (dynamic call graph) data
-	err = os.New().RemoveAll(dcgDataDirPath())
+	err = osfs.New().RemoveAll(dcgDataDirPath())
 	if nil != err {
 		panic(fmt.Errorf("unable to cleanup DCG (dynamic call graph) data at path: %v\n", dcgDataDirPath()))
 	}