+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file in
 accordance with [keepachangelog.com](http://keepachangelog.com/)
 
+## 0.1.18 - 2017-03-28
+
+### Changed
+
+- [Don't recreate node on self-update](https://github.com/opctl/opctl/issues/169)
+
+### Fixed
+
+- [Multiple opctl networks created leading to lack of inter-container connectivity](https://github.com/opctl/opctl/issues/167)
+
+## 0.1.16 - 2017-03-26
+
+### Fixed
+
+- [Outputs internal to op call graph not initialized](https://github.com/opctl/opctl/issues/165)
+
 ## 0.1.15 - 2017-03-23
 
 ### Added